server: Introduce new runnnig job tracker
This introduces an iteractive tracker which iterates through the list of
running jobs and see if there is same run_id (ie. multi config/sets), then
have a common counter for the job. Useful for time-estimator in big picture.

In the current logic, the counter will reset if a set is completed and recomputes.

Strictly speaking it is really running job tracker.