metrics_gather: Migrate to Y4M encode pipeline for VVC-VTM
We have to exclude chroma-format 420mpeg2 due to bug in upstream,

Decoding y4ms do not retain frame-rate of input. Due to some SEI storage
reasoning, so currently best to decode to YUV, and then convert to Y4M
If decoding to Y4M from bitstream is needed, we need to have
`--HrdParametersPresent=1 --SEIPictureTiming=1 --SEIBufferingPeriod=1 --RCCpbSize=1000`
as parameters.